requirements.txt
google-generativeai>=0.8.0
langchain-google-genai>=0.0.7
langchain-core>=0.1.18
langchain>=0.1.3
langgraph>=0.0.33,<0.1.0
requests>=2.31.0
beautifulsoup4>=4.12.2
streamlit>=1.31.0
streamlit_option_menu>=0.3.6
pdfplumber>=0.10.3
docx2txt>=0.8
python-dotenv>=1.0.0